+++ src/interpreters/GFM_original.rs
@@ -93,7 +93,8 @@ impl GFM_original {
             "matlab" => "matlab",
             "objectivec" => "objcpp",
             "swift" => "swift",
-            &_ => &self.default_filetype,
+            "" => &self.default_filetype,
+            a => a,
         }
         .to_string()
     }
@@ -162,7 +163,7 @@ impl Interpreter for GFM_original {
     }
 
     fn get_name() -> String {
-        String::from("Markdown_original")
+        String::from("GFM_original")
     }
 
     fn get_current_level(&self) -> SupportLevel {
@@ -243,3 +244,26 @@ impl Interpreter for GFM_original {
         )));
     }
 }
+
+
+#[cfg(test)]
+mod test_GFM_original {
+    use super::*;
+
+    #[test]
+    //actually safe to run in parrallel because GFM modify sniprun root and thus 
+    //where the underlying interpreter puts its files
+    fn simple_bloc(){
+        let mut data = DataHolder::new();
+        data.current_bloc = String::from("print(3)");
+        data.filepath = String::from("ressources/markdown.md");
+        data.filetype = String::from("python");
+        data.range = [1,3];
+        
+        let mut interpreter = GFM_original::new(data);
+        let res = interpreter.execute();
+        let string_result = res.unwrap();
+        assert_eq!(string_result, "3\n");
+    }
+}
+