+++ src/interpreters/Python3_original.rs
@@ -155,7 +155,7 @@ impl Interpreter for Python3_original {
         if !self.imports.is_empty() {
             let mut indented_imports = String::new();
             for import in self.imports.lines() {
-                indented_imports = indented_imports + "\t" + import+"\n";
+                indented_imports = indented_imports + "\t" + import + "\n";
             }
 
             self.imports = String::from("\ntry:\n") + &indented_imports + "\nexcept:\n\tpass\n";
@@ -237,7 +237,6 @@ impl ReplLikeInterpreter for Python3_original {
 
         Ok(())
     }
-
 }
 
 #[cfg(test)]
@@ -248,6 +247,7 @@ mod test_python3_original {
     fn run_all() {
         simple_print();
         print_quote();
+        get_import();
     }
     fn simple_print() {
         let mut data = DataHolder::new();
@@ -259,7 +259,7 @@ mod test_python3_original {
         let string_result = res.unwrap();
         assert_eq!(string_result, "lol 1\n");
     }
-    fn print_quote(){
+    fn print_quote() {
         let mut data = DataHolder::new();
         data.current_bloc = String::from("print(\"->\\\"\",1);");
         let mut interpreter = Python3_original::new(data);
@@ -269,4 +269,24 @@ mod test_python3_original {
         let string_result = res.unwrap();
         assert_eq!(string_result, "->\" 1\n");
     }
+
+    fn get_import(){
+        let mut data = DataHolder::new();
+        data.current_bloc = String::from("print(cos(0))");
+
+        data.filepath = String::from("ressources/import2.py");
+        let dfpc = data.filepath.clone();
+        let mut file = File::create(&data.filepath).unwrap();
+        file.write_all(b"from math import cos").unwrap();
+
+        let mut interpreter = Python3_original::new(data);
+        let res = interpreter.run_at_level(SupportLevel::Import);
+
+        // should panic if not an Ok()
+        let string_result = res.unwrap();
+        assert_eq!(string_result, "1.0\n");
+
+        std::fs::remove_file(dfpc).unwrap();
+    }
+
 }